move xml serialization declaration parser code to its own source file
