+++ tools/code_generators/luagamestate_support_gen/CMakeLists.txt
@@ -1,4 +1,3 @@
-
 # project name, and type
 PROJECT(OpenApoc_LuaGamestateSupportGen CXX C)
 
@@ -8,9 +7,14 @@ CMAKE_MINIMUM_REQUIRED(VERSION 3.1)
 find_package(Boost REQUIRED COMPONENTS program_options)
 
 set (LUAGAMESTATESUPPORTGEN_SOURCE_FILES
-	main.cpp)
+	main.cpp ../serialization_xml.cpp)
 
-set (GAMESTATESERIALIZEGEN_HEADER_FILES)
+source_group(luagamestatesupportgen\\sources FILES ${LUAGAMESTATESUPPORTGEN_SOURCE_FILES})
+
+set (GAMESTATESERIALIZEGEN_HEADER_FILES
+	../serialization_xml.h)
+
+source_group(luagamestatesupportgen\\headers FILES ${LUAGAMESTATESUPPORTGEN_HEADER_FILES})
 
 list(APPEND ALL_SOURCE_FILES ${LUAGAMESTATESUPPORTGEN_SOURCE_FILES})
 list(APPEND ALL_HEADER_FILES ${LUAGAMESTATESUPPORTGEN_HEADER_FILES})
@@ -25,3 +29,4 @@ target_link_libraries(OpenApoc_LuaGamestateSupportGen OpenApoc_LibPugixml
 
 target_include_directories(OpenApoc_LuaGamestateSupportGen PUBLIC ${Boost_INCLUDE_DIRS})
 target_include_directories(OpenApoc_LuaGamestateSupportGen PUBLIC ${CMAKE_SOURCE_DIR})
+target_include_directories(OpenApoc_LuaGamestateSupportGen PUBLIC ..)
