+++ tools/code_generators/luagamestate_support_gen/luagamestate_support_gen.vcxproj
@@ -1,4 +1,4 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
+<?xml version="1.0" encoding="utf-8"?>
 <Project DefaultTargets="Build" ToolsVersion="14.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
   <ItemGroup Label="ProjectConfigurations">
     <ProjectConfiguration Include="Debug|Win32">
@@ -19,20 +19,22 @@
     </ProjectConfiguration>
   </ItemGroup>
   <ItemGroup>
-    <ClCompile Include="..\..\dependencies\pugixml\src\pugixml.cpp" />
+    <ClCompile Include="..\..\..\dependencies\pugixml\src\pugixml.cpp" />
+    <ClCompile Include="..\serialization_xml.cpp" />
     <ClCompile Include="main.cpp" />
   </ItemGroup>
   <ItemGroup>
-    <ClInclude Include="..\..\dependencies\pugixml\src\pugiconfig.hpp" />
-    <ClInclude Include="..\..\dependencies\pugixml\src\pugixml.hpp" />
+    <ClInclude Include="..\..\..\dependencies\pugixml\src\pugiconfig.hpp" />
+    <ClInclude Include="..\serialization_xml.h" />
+    <ClInclude Include="..\..\..\dependencies\pugixml\src\pugixml.hpp" />
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
   </ItemGroup>
   <PropertyGroup Label="Globals">
-    <ProjectGuid>{31FBCCBD-BFD2-45FC-B9DF-3375684A3157}</ProjectGuid>
+    <ProjectGuid>{31FBCCBD-BFD2-45FC-B9DF-3375684A3158}</ProjectGuid>
     <Keyword>Win32Proj</Keyword>
-    <RootNamespace>gamestate_serialize_gen</RootNamespace>
+    <RootNamespace>luagamestate_support_gen</RootNamespace>
     <WindowsTargetPlatformVersion>8.1</WindowsTargetPlatformVersion>
   </PropertyGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.Default.props" />
